separated download code in its own file
